# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Imagesystems",
    "description": "Imagesystems \u2014 hotel & real estate image enhancement, generation, and batch pipelines",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/image-generation/compose": {
      "post": {
        "tags": [
          "image-generation"
        ],
        "summary": "Compose Prompt",
        "description": "Phase 1a \u2014 Software intelligence: turn a vague request into a concrete image prompt.\nDoes not call the image API or consume image-generation quota.",
        "operationId": "compose_prompt_api_image_generation_compose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationComposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerationComposeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/image-generation/generate": {
      "post": {
        "tags": [
          "image-generation"
        ],
        "summary": "Generate Image",
        "description": "Interpret (optional) \u2192 image model \u2192 persist as a new Image (original), same as an upload for Operations.",
        "operationId": "generate_image_api_image_generation_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerationGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/knowledge/scenarios": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_api_knowledge_scenarios_get",
        "parameters": [
          {
            "name": "vertical",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact vertical slug.",
              "title": "Vertical"
            },
            "description": "Exact vertical slug."
          },
          {
            "name": "room_type_hint",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ROOM_CONTEXTS key.",
              "title": "Room Type Hint"
            },
            "description": "ROOM_CONTEXTS key."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Derived risk bucket (keyword match). See /scenarios/filters \u2192 risk_categories.",
              "title": "Category"
            },
            "description": "Derived risk bucket (keyword match). See /scenarios/filters \u2192 risk_categories."
          },
          {
            "name": "rule_ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scenario must reference this policy id.",
              "title": "Rule Ref"
            },
            "description": "Scenario must reference this policy id."
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact scenario id.",
              "title": "Scenario Id"
            },
            "description": "Exact scenario id."
          },
          {
            "name": "id_prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scenario id starts with this prefix.",
              "title": "Id Prefix"
            },
            "description": "Scenario id starts with this prefix."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring over narrative + id + vertical.",
              "title": "Q"
            },
            "description": "Case-insensitive substring over narrative + id + vertical."
          },
          {
            "name": "min_depth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 20,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum scenario_depth_score.",
              "title": "Min Depth"
            },
            "description": "Minimum scenario_depth_score."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Max items to return after offset.",
              "title": "Limit"
            },
            "description": "Max items to return after offset."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "description": "Skip this many filtered rows.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Skip this many filtered rows."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenariosListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Scenario",
        "operationId": "get_scenario_api_knowledge_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnhancementScenarioItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnhancementScenarioItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vertical": {
            "type": "string",
            "title": "Vertical"
          },
          "room_type_hint": {
            "type": "string",
            "title": "Room Type Hint"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "setup": {
            "type": "string",
            "title": "Setup"
          },
          "forbidden_model_behaviors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Model Behaviors"
          },
          "allowed_model_behaviors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Model Behaviors"
          },
          "rule_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Refs"
          },
          "risk_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Categories",
            "description": "Keyword-derived risk buckets (see GET /api/knowledge/scenarios/filters)."
          },
          "depth_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Depth Score",
            "description": "Heuristic richness score for the scenario narrative."
          },
          "rule_summaries": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Rule Summaries",
            "description": "rule_ref id \u2192 short policy line from the server registry."
          }
        },
        "type": "object",
        "required": [
          "id",
          "vertical",
          "room_type_hint",
          "title",
          "setup",
          "forbidden_model_behaviors",
          "allowed_model_behaviors",
          "rule_refs",
          "depth_score"
        ],
        "title": "EnhancementScenarioItem",
        "description": "One catalog row plus derived intelligence for clients."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageGenerationComposeRequest": {
        "properties": {
          "user_request": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 3,
            "title": "User Request"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "gemini"
            ],
            "title": "Provider",
            "description": "Which provider's key to use for the interpretation LLM (match the key you use to generate).",
            "default": "openai"
          }
        },
        "type": "object",
        "required": [
          "user_request"
        ],
        "title": "ImageGenerationComposeRequest",
        "description": "Turn natural language into a concrete image prompt (preview before generation)."
      },
      "ImageGenerationComposeResponse": {
        "properties": {
          "interpreted_prompt": {
            "type": "string",
            "title": "Interpreted Prompt",
            "description": "Refined prompt ready for the image model."
          },
          "short_title": {
            "type": "string",
            "title": "Short Title",
            "description": "Short label for filenames and UI."
          }
        },
        "type": "object",
        "required": [
          "interpreted_prompt",
          "short_title"
        ],
        "title": "ImageGenerationComposeResponse"
      },
      "ImageGenerationGenerateRequest": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Description",
            "description": "Natural language (interpret=True) or exact prompt (interpret=False)."
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "gemini"
            ],
            "title": "Provider"
          },
          "interpret": {
            "type": "boolean",
            "title": "Interpret",
            "description": "When True, run an LLM step to expand the description into a detailed image prompt.",
            "default": true
          },
          "model": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Model"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "description": "Image quality: low | medium | high \u2014 passed to OpenAI; same tier is applied to Gemini (prompt + encoder settings) for matching behavior.",
            "default": "high"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format",
            "description": "png, jpeg, or webp \u2014 OpenAI returns that format natively; Gemini output is re-encoded to the same format and file extension.",
            "default": "png"
          },
          "run_enhancement_pipeline": {
            "type": "boolean",
            "title": "Run Enhancement Pipeline",
            "description": "Phase 2: run full enhancement after generation \u2014 not implemented yet.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "description",
          "provider",
          "model"
        ],
        "title": "ImageGenerationGenerateRequest",
        "description": "Generate a new raster from text; optionally interpret vague text first."
      },
      "ImageGenerationGenerateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ImageVersionResponse"
            },
            "type": "array",
            "title": "Versions"
          },
          "resolved_prompt": {
            "type": "string",
            "title": "Resolved Prompt",
            "description": "Exact prompt sent to the image API."
          },
          "used_interpretation": {
            "type": "boolean",
            "title": "Used Interpretation",
            "description": "Whether the LLM interpretation step ran."
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "created_at",
          "resolved_prompt",
          "used_interpretation"
        ],
        "title": "ImageGenerationGenerateResponse",
        "description": "New Image row plus generation metadata."
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "result_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Version Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "job_type",
          "status",
          "progress_pct",
          "created_at"
        ],
        "title": "JobResponse"
      },
...
      "ScenariosListResponse": {
        "properties": {
          "catalog_version": {
            "type": "integer",
            "title": "Catalog Version"
          },
          "catalog_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Description"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Scenarios in the full catalog before filters."
          },
          "filtered": {
            "type": "integer",
            "title": "Filtered",
            "description": "Count after filters, before limit/offset slice."
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EnhancementScenarioItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "catalog_version",
          "total",
          "filtered",
          "items"
        ],
        "title": "ScenariosListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}